% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregated_fetal_network}
\alias{aggregated_fetal_network}
\title{Aggregated fetal co-expression network}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20709 rows and 20709 columns.
}
\usage{
aggregated_fetal_network
}
\description{
This is the aggregated fetal network from the meta-analysis perform by Werner and Gillis (citation).
This network is derived from single-cell RNA sequencing expression data of developing fetal brain tissue ranging from gestational week 5 to 25
}
\keyword{datasets}
