% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BP_GO_mat}
\alias{BP_GO_mat}
\title{Biological Process go terms}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20700 rows and 15960 columns.
}
\usage{
BP_GO_mat
}
\description{
This is a binary martix detailing the gene-GOterm relationships for the Biological Process GO ontology
}
\keyword{datasets}
