% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CC_GO_mat}
\alias{CC_GO_mat}
\title{Cellular Component go terms}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20709 rows and 2009 columns.
}
\usage{
CC_GO_mat
}
\description{
This is a binary matrix detailing the gene-GOterm relationships for the Cellular Component GO ontology
}
\keyword{datasets}
