% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserved_coexp.R
\name{compute_auc}
\alias{compute_auc}
\title{Compute an auroc statistic using the Mann Whitney U test}
\usage{
compute_auc(scores, label)
}
\arguments{
\item{scores}{some metric for ranking data, must be compatible with the rank() function}

\item{label}{a binary numeric vector indicating data belonging to positives (1's) or negatives (0's)}
}
\value{
an auroc statistic
}
\description{
Compute an auroc statistic using the Mann Whitney U test
}
