% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserved_coexp.R
\name{ecdf_fun}
\alias{ecdf_fun}
\title{Compute a percentile from ecdf}
\usage{
ecdf_fun(x, perc)
}
\arguments{
\item{x}{distribution to compute ecdf}

\item{perc}{percentile to report}
}
\value{
double
}
\description{
Compute a percentile from ecdf
}
