% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fetal_meta_markers}
\alias{fetal_meta_markers}
\title{Fetal metaMarkers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 135472 rows and 50 columns.
}
\usage{
fetal_meta_markers
}
\description{
This is a data frame of fetal MetaMarkers as computed in Werner and Gillis (citation). These cell-type markers were derived from annotated
fetal brain scRNA-seq data spanning gestational week 5 - 25.
}
\keyword{datasets}
