% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_GO.R
\name{fit_to_GO}
\alias{fit_to_GO}
\title{Fit an expression matrix with genes as rows to the GO universe of GO annotations}
\usage{
fit_to_GO(dataset)
}
\arguments{
\item{dataset}{a matrix with genes as rows, samples (cells) as columns, and expression counts as data, we recommend a CPM normalized expression matrix}
}
\value{
An expression matrix zero-padded for the missing GO genes in the original expression matrix
}
\description{
Fit an expression matrix with genes as rows to the GO universe of GO annotations
}
