% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserved_coexp.R
\name{get_geneset_preserved_score}
\alias{get_geneset_preserved_score}
\title{Compute the preserved co-expression score for a geneset}
\usage{
get_geneset_preserved_score(
  test_network,
  geneset,
  num_top = 10,
  parallel = F,
  cores = 10
)
}
\arguments{
\item{test_network}{the test co-expression network, we recommend using the rank standardized networks}

\item{geneset}{vector of strings}

\item{num_top}{integer, the number of top co-expressed partners of the test_gene to compute the auroc over}

\item{parallel}{boolean, if the parallelized version should be implemented, requires the parallel R package}

\item{cores}{integer, the number of computer cores for the parallelized version}
}
\value{
integer, the mean gene preserved co-expression auroc for the geneset
}
\description{
Compute the preserved co-expression score for a geneset
}
