% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserved_coexp.R
\name{get_GO_term_scores}
\alias{get_GO_term_scores}
\title{Get a p-value for all GO terms quantifying whether each GO term has less (left-sided p-value) or more (right-sided p-value) preserved co-expression than expected by the length of the GO term}
\usage{
get_GO_term_scores(
  aggregated_fetal_network,
  test_network,
  go_matrix,
  parallel = F
)
}
\arguments{
\item{aggregated_fetal_network}{the aggregated fetal co-expression network to act as a reference network. Any co-expression network will work, with the same gene annotations}

\item{test_network}{the test co-expression network, we recommend using the rank standardized networks}

\item{go_matrix}{one of the 3 binary matrices provided encoding the mappings between all genes and their GO terms
#'@param parallel boolean, if the parallelized version should be implemented, requires the parallel R package}
}
\value{
data.frame of the resulting statistics per GO term
}
\description{
Get a p-value for all GO terms quantifying whether each GO term has less (left-sided p-value) or more (right-sided p-value) preserved co-expression than expected by the length of the GO term
}
