% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserved_coexp.R
\name{get_metaMarker_preserved_coexp}
\alias{get_metaMarker_preserved_coexp}
\title{Compute the preserved co-expression of a cell-type marker gene set from our list of fetal MetaMarkers}
\usage{
get_metaMarker_preserved_coexp(
  aggregated_fetal_network,
  test_network,
  metamarkers,
  celltype,
  num_markers,
  parallel = F,
  cores = 10
)
}
\arguments{
\item{aggregated_fetal_network}{the aggregated fetal co-expression network to act as a reference network. Any co-expression network will work, with the same gene annotations}

\item{test_network}{the test co-expression network, we recommend using the rank standardized networks}

\item{metamarkers}{A datatable of metamarkers, included as data for this package}

\item{celltype}{string belonging to one of the celltypes included in the MetaMarkers. Must be one of c('Dividing_Progenitor','Neural_Progenitor','Intermediate_Progenitor','GABAergic','Glutamatergic','Non-neuronal')}

\item{num_markers}{int, number of MetaMarkers, ranked by recurrent differential expression}

\item{parallel}{boolean, if the parallelized version should be implemented, requires the parallel R package}

\item{cores}{integer, the number of computer cores for the parallelized version}
}
\value{
a named vector of gene preserved co-expression aurocs
}
\description{
Compute the preserved co-expression of a cell-type marker gene set from our list of fetal MetaMarkers
}
