% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_genes}
\alias{go_genes}
\title{List of GO genes}
\format{
An object of class \code{character} of length 20700.
}
\usage{
go_genes
}
\description{
List of all genes in the GO universe
}
\keyword{datasets}
