% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserved_coexp.R
\name{load_fetal_coexp}
\alias{load_fetal_coexp}
\title{Load the aggregated fetal co-expression network as computed in Werner and Gillis (citation)}
\usage{
load_fetal_coexp(
  data_url =
    "https://labshare.cshl.edu/shares/gillislab/resource/preserved_fetal_coexp/ranked_agg_fetal_coexp_network_8_17_23.Rdata"
)
}
\arguments{
\item{data_url}{url for the aggregated fetal co-expression network hosted on the Gillis lab public FTP server}
}
\value{
matrix
}
\description{
Load the aggregated fetal co-expression network as computed in Werner and Gillis (citation)
}
