% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_presCoexp_df}
\alias{meta_presCoexp_df}
\title{Fetal and organoid Meta-analysis preserved co-expression results}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1568 rows and 4 columns.
}
\usage{
meta_presCoexp_df
}
\description{
dataframe containing the cell-type preserved co-expression scores for the fetal and organoid datasets analyzed in Werner and Gillis (citation)
}
\keyword{datasets}
