% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MF_GO_mat}
\alias{MF_GO_mat}
\title{Molecular Function go terms}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20709 rows and 4878 columns.
}
\usage{
MF_GO_mat
}
\description{
This is a binary matrix detailing the gene-GOterm relationships for the Molecular Function GO ontology
}
\keyword{datasets}
