% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\docType{data}
\name{np}
\alias{np}
\title{Imports the python numpy library using reticulate}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
np
}
\description{
Imports the python numpy library using reticulate
}
\keyword{datasets}
