% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserved_coexp.R
\name{plot_metafetal_results}
\alias{plot_metafetal_results}
\title{Plot the preserved co-expression scores for the top 100 Fetal MetaMarkers, placing the user's dataset in reference to our meta-analysis of fetal datasets}
\usage{
plot_metafetal_results(user_network)
}
\arguments{
\item{user_network}{gene x gene co-expression network of the user's data, we recommend using the rank standardized network}
}
\value{
ggplot object
}
\description{
Plot the preserved co-expression scores for the top 100 Fetal MetaMarkers, placing the user's dataset in reference to our meta-analysis of fetal datasets
}
