% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preservedCoexp-package.R
\docType{package}
\name{preservedCoexp-package}
\alias{preservedCoexp}
\alias{preservedCoexp-package}
\title{preservedCoexp: Compute the preserved co-expression of genes between two co-expression networks}
\description{
This package allows users to explore the preserved co-expression of genes from their single-cell RNA-seq data using a fetal brain reference co-expression network. It provides a meta-analytic co-expression network derived from fetal brain single-cell RNAseq datasets and functions for users to generate co-expression networks of their own single-cell RNAseq datasets. Users can quantify the strength of preserved co-expression between their data and our meta-analytic co-expression network for individual genes or genesets. We provide functions to quantify the preserved co-expression of our meta-analytically derived fetal cell-type markers as well as GO terms.
}
\author{
\strong{Maintainer}: Jonathan Werner \email{werner@cshl.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
