% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{rank_coexpression}
\alias{rank_coexpression}
\title{Generate a rank-standardized co-expression matrix}
\usage{
rank_coexpression(corr_matrix)
}
\arguments{
\item{corr_matrix}{a gene x gene correlation matrix, as output from get_spearman()}
}
\value{
a gene x gene correlation matrix rank standardized between 0 and 1
}
\description{
Generate a rank-standardized co-expression matrix
}
